Moved developer tools to the settings tab
At request of Geo
