+++ components/DashboardPage.vue
@@ -17,10 +17,6 @@
             <ModerationIcon />
             Moderation
           </nuxt-link>
-          <nuxt-link :to="'/dashboard/developer'" class="tab last">
-            <DevTools />
-            Developer Tools
-          </nuxt-link>
           <nuxt-link :to="'/dashboard/settings'" class="tab last">
             <SettingsIcon />
             Settings
@@ -41,7 +37,6 @@
 import ModIcon from '~/assets/images/sidebar/mod.svg?inline'
 import ModerationIcon from '~/assets/images/sidebar/admin.svg?inline'
 import SettingsIcon from '~/assets/images/sidebar/settings.svg?inline'
-import DevTools from '~/assets/images/sidebar/devtools.svg?inline'
 
 export default {
   name: 'DashboardPage',
@@ -49,7 +44,6 @@ export default {
     ModIcon,
     ModerationIcon,
     SettingsIcon,
-    DevTools,
   },
 }
 </script>
